updated list
Fixed a typo, made list alphabetical, added litElement and added statement about where this list comes from and how to contribute to it. Also removed x-tag as it is not ES modules compatible (at this time) and added statements about the importance of selecting a library which adopts all 4 of these specs.